requirements.txt
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
Flask-Login
flask-migrate